+++ sdk/utils/parse.js
@@ -1,7 +1,12 @@
 
 const parse = require('starkcore').parse;
-const {apiVersion, sdkVersion, host, language, timeout } = require('../utils/rest.js');
+const starkBank = require('../../index.js');
+const starkHost = require('starkcore').starkHost;
 
+const apiVersion = starkBank.apiVersion;
+const sdkVersion = starkBank.version;
+const host = starkHost.bank;
+const timeout = starkBank.timeout;
 
 exports.parseObjects = function (objects, resource, resourceClass) {
     return parse.parseObjects(
@@ -11,7 +16,7 @@ exports.parseObjects = function (objects, resource, resourceClass) {
     );
 }
 
-exports.parseAndVerify = async function (resource, content, signature, user = null) {
+exports.parseAndVerify = async function (resource, content, signature, user = starkBank.user) {
     return parse.parseAndVerify(
         resource,
         content,
@@ -20,12 +25,12 @@ exports.parseAndVerify = async function (resource, content, signature, user = nu
         apiVersion,
         host,
         user,
-        language,
+        starkBank.language,
         timeout
     );
 }
 
-exports.verify = async function (content, signature, user = null) {
+exports.verify = async function (content, signature, user = starkBank.user) {
     return parse.verify (
         content,
         signature,
@@ -33,7 +38,7 @@ exports.verify = async function (content, signature, user = null) {
         apiVersion,
         host,
         user,
-        language,
+        starkBank.language,
         timeout
     );
 }